requirements.txt
arrow==0.6.0
Chameleon==2.22
click==4.1
ipdb==0.8.1
ipython==3.2.1
Jinja2==2.8
Mako==1.0.1
MarkupSafe==0.23
PasteDeploy==1.5.2
psycopg2==2.6.1
Pygments==2.0.2
pyramid==1.5.7
pyramid-chameleon==0.3
pyramid-debugtoolbar==2.4
pyramid-jinja2==2.5
pyramid-mako==1.0.2
pyramid-redis-sessions==1.0.1
pyramid-tm==0.12
python-dateutil==2.4.2
python-dotenv==0.1.2
redis==2.10.3
repoze.lru==0.6
six==1.9.0
SQLAlchemy==1.0.8
SQLAlchemy-Utils==0.30.15
transaction==1.4.4
translationstring==1.3
venusian==1.0
waitress==0.8.9
WebOb==1.4.1
wheel==0.24.0
zope.deprecation==4.1.2
zope.interface==4.1.2
zope.sqlalchemy==0.7.6
